Move EtcdStorage to a separate module
