+++ asto-core/pom.xml
@@ -77,19 +77,6 @@ SOFTWARE.
       <artifactId>resilience4j-retry</artifactId>
       <version>1.7.1</version>
     </dependency>
-    <!-- etcd client -->
-    <dependency>
-      <groupId>io.etcd</groupId>
-      <artifactId>jetcd-core</artifactId>
-      <version>0.7.1</version>
-      <optional>true</optional>
-      <exclusions>
-        <exclusion>
-          <groupId>com.google.android</groupId>
-          <artifactId>annotations</artifactId>
-        </exclusion>
-      </exclusions>
-    </dependency>
     <dependency>
       <groupId>org.junit.jupiter</groupId>
       <artifactId>junit-jupiter-params</artifactId>
