--- asto-s3/src/test/java/com/artipie/asto/factory/package-info.java
@@ -1,11 +1,0 @@
-/*
- * The MIT License (MIT) Copyright (c) 2020-2022 artipie.com
- * https://github.com/artipie/asto/LICENSE.txt
- */
-
-/**
- * Tests fo S3 storage factory classes.
- *
- * @since 0.1
- */
-package com.artipie.asto.factory;
