+++ pom.xml
@@ -84,6 +84,7 @@ SOFTWARE.
     <module>asto-redis</module>
     <module>asto-s3</module>
     <module>asto-vertx-file</module>
+    <module>asto-etcd</module>
   </modules>
   <properties>
     <qulice.license>${project.basedir}/../LICENSE.header</qulice.license>
